perf(main): :zap: Improved game replay method
I improved the code in how it replays a game calling the play() method rather than the main() method as they would not need the instructions again. If they do they could refer to it above. I also improved the colouring in the code for the prompts, pc text, etc.